backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
numpy==1.26.3
sounddevice==0.4.6
scipy==1.11.4
openai-whisper==20231117
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2
ffmpeg-python==0.2.0
